requirements.txt
twilio>=9,<10
uvicorn>=0.24,<0.25
fastapi>=0.100,<0.101
requests>=2.31,<3
jupyterlab>=4,<5
openai>=1.30,<2
sounddevice>=0.4,<0.5
soundfile>=0.12,<0.13
numpy>=1.26,<2.0

# Optional dev/debug tools (uncomment if needed)
python-dotenv>=1.0,<2
